remove highlighted area when using buffer
