+++ apps/fishing-map/features/sidebar/SidebarHeader.tsx
@@ -54,7 +54,6 @@ import { HOME, REPORT, ROUTE_TYPES, WORKSPACE } from 'routes/routes'
 import { EMPTY_FILTERS, IMO_LENGTH, SSVID_LENGTH, SearchType } from 'features/search/search.config'
 import { resetAreaDetail } from 'features/areas/areas.slice'
 import { selectReportAreaIds } from 'features/reports/reports.selectors'
-import { QueryParams } from 'types'
 import { useSearchFiltersConnect } from 'features/search/search.hook'
 import { resetVesselState } from 'features/vessel/vessel.slice'
 import { cleanVesselSearchResults } from 'features/search/search.slice'
@@ -65,7 +64,7 @@ import { isPrivateWorkspaceNotAllowed } from 'features/workspace/workspace.utils
 import { selectUserData } from 'features/user/selectors/user.selectors'
 import { setModalOpen } from 'features/modals/modals.slice'
 import TooltipContainer from 'features/workspace/shared/TooltipContainer'
-import { useHighlightReportArea } from 'features/map/popups/categories/ContextLayers.hooks'
+import { useHighlightReportArea } from 'features/reports/reports.hooks'
 import { useClipboardNotification } from './sidebar.hooks'
 import styles from './SidebarHeader.module.css'
